+++ widgets/WidgetScrollArea.cpp
@@ -2,7 +2,13 @@
 #include <QScrollBar>
 #include <QEvent>
 
-QSize WidgetScrollArea::sizeHint() const {
+WidgetScrollArea::WidgetScrollArea(QWidget *parent) :
+  QScrollArea(parent),
+  loggingCategory("WidgetScrollArea")
+{}
+
+QSize WidgetScrollArea::sizeHint() const
+{
   return QSize(
     this->viewportSizeHint().width() +
       (this->verticalScrollBar()->isVisible() ? this->verticalScrollBar()->width() : 0),
@@ -10,8 +16,8 @@ QSize WidgetScrollArea::sizeHint() const {
       (this->horizontalScrollBar()->isVisible() ? this->horizontalScrollBar()->height() : 0)
   );
 }
-
-QSize WidgetScrollArea::minimumSizeHint() const {
+QSize WidgetScrollArea::minimumSizeHint() const
+{
   return QSize(0, 0);
 }
 
@@ -19,6 +25,10 @@ bool WidgetScrollArea::event(QEvent *event)
 {
   auto result = QScrollArea::event(event);
   if (event->type() == QEvent::LayoutRequest) {
+    qCDebug(loggingCategory)
+      << "Updating geometry,"
+      << "sizeHint" << sizeHint()
+      << "widget.sizeHint" << (widget() ? widget()->sizeHint() : QSize());
     updateGeometry(); // makes this scrollarea resize based on new size hint
   }
   return result;