+++ widgets/WidgetScrollArea.h
@@ -1,17 +1,21 @@
 #pragma once
 
 #include <QScrollArea>
+#include <QLoggingCategory>
 
 class WidgetScrollArea : public QScrollArea
 {
   Q_OBJECT
 
 public:
-  using QScrollArea::QScrollArea;
+  WidgetScrollArea(QWidget *parent = nullptr);
 
   virtual QSize sizeHint() const;
   virtual QSize minimumSizeHint() const;
 
 protected:
   bool event(QEvent *event) override;
+
+private:
+  QLoggingCategory loggingCategory;
 };