+++ changelog.md
@@ -8,4 +8,5 @@
   - optimize retrieving table row rect when rendering checked rows
   - fix inelegant workarounds forcing updates/recalculations/repaints
   - reimplement scroll area and scroll bar for content and menu by using CSS for styling and extending `QScrollArea` and `QScrollbar` for calculation adjustments to support all scrolling features (by mouse wheel and touchpad)
+  - make table cells size adaptive based on the contents
   - add `EventCallback` internal tool for future